Cruds 43 79 - ACF

+++ backend/controllers/EventoController.php
@@ -4,6 +4,8 @@ namespace backend\controllers;
 
 use common\models\Evento;
 use common\models\EventoSearch;
+use yii\web\Response;
+use yii\filters\AccessControl;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
@@ -21,13 +23,26 @@ class EventoController extends Controller
         return array_merge(
             parent::behaviors(),
             [
+                'access' => [
+                    'class' => AccessControl::class,
+                    'rules' => [
+                        [
+                            'actions' => ['index', 'view', 'update', 'delete'],
+                            'allow' => true,
+                            'roles' => ['admin'],
+                        ],
+                    ],
+                    'denyCallback' => function ($rule, $action) {
+                        return $this->redirect(['site/index']);
+                    }
+                ],
                 'verbs' => [
                     'class' => VerbFilter::className(),
                     'actions' => [
                         'delete' => ['POST'],
                     ],
                 ],
-            ]
+            ],
         );
     }
 
@@ -51,11 +66,15 @@ class EventoController extends Controller
      * Displays a single Evento model.
      * @param int $id ID
      * @param int $carta_id Carta ID
-     * @return string
+     * @return string|Response
      * @throws NotFoundHttpException if the model cannot be found
      */
     public function actionView($id, $carta_id)
     {
+        if (!\Yii::$app->user->can('readEvento')) {
+            return $this->redirect(['site/index']);
+        }
+
         return $this->render('view', [
             'model' => $this->findModel($id, $carta_id),
         ]);
@@ -93,6 +112,10 @@ class EventoController extends Controller
      */
     public function actionUpdate($id, $carta_id)
     {
+        if (!\Yii::$app->user->can('updateEvento')) {
+            return $this->redirect(['site/index']);
+        }
+
         $model = $this->findModel($id, $carta_id);
 
         if ($this->request->isPost && $model->load($this->request->post()) && $model->save()) {
@@ -114,6 +137,10 @@ class EventoController extends Controller
      */
     public function actionDelete($id, $carta_id)
     {
+        if (!\Yii::$app->user->can('deleteEvento')) {
+            return $this->redirect(['site/index']);
+        }
+
         $this->findModel($id, $carta_id)->delete();
 
         return $this->redirect(['index']);
